Revert "Fix Trainee Page Modal (#43)"
This reverts commit 483b49e5617cb313850b2d0b42c0b1a0db27d349.
